+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### New
 - adding support for S3 to store bundles
+- adding seedfarmer version tag to toolchain and deployment roles 
 
 ### Changes
 - adding local path of manifests that fail to load to the actual final string printed